+++ agb/src/display/font/sprite.rs
@@ -39,6 +39,7 @@ mod tests {
 
     use crate::{
         display::{
+            Rgb15,
             font::{AlignmentKind, ChangeColour, Font, Layout},
             palette16::Palette16,
         },
@@ -53,9 +54,9 @@ mod tests {
         let mut gfx = gba.graphics.get();
 
         static PALETTE: Palette16 = const {
-            let mut palette = [0x0; 16];
-            palette[1] = 0xFF_FF;
-            palette[2] = 0x10_7C;
+            let mut palette = [Rgb15::BLACK; 16];
+            palette[1] = Rgb15::WHITE;
+            palette[2] = Rgb15(0x10_7C);
             Palette16::new(palette)
         };
 
@@ -91,8 +92,9 @@ mod tests {
         let mut gfx = gba.graphics.get();
 
         static PALETTE: Palette16 = const {
-            let mut palette = [0x0; 16];
-            palette[1] = 0xFF_FF;
+            let mut palette = [Rgb15::BLACK; 16];
+            palette[1] = Rgb15::WHITE;
+            palette[2] = Rgb15(0x10_7C);
             Palette16::new(palette)
         };